requirements.txt
dj-database-url==0.4.2
Django==2.0.2
django-filter==1.1.0
django-widget-tweaks==1.4.1
python-decouple==3.1
